refactor(queries): final lib/queries split — extract OLS + protocol fees + LP positions
Final slice of the lib/queries split. After PR #277 (pools) and PR #278
(config), the remaining 7 queries move to per-domain modules:

- lib/queries/lp.ts        — POOL_LP_POSITIONS, UNIQUE_LP_ADDRESSES
- lib/queries/ols.ts       — OLS_POOL, OLS_LIQUIDITY_EVENTS_PAGE,
                             OLS_LIQUIDITY_EVENTS_COUNT, ALL_OLS_POOLS
- lib/queries/protocol.ts  — PROTOCOL_FEE_TRANSFERS_ALL

queries.ts is now a pure barrel re-exporting all five per-domain modules
(pools, config, lp, ols, protocol) so existing `from "@/lib/queries"`
imports stay stable; consumers may also import directly from the per-
domain files.

Drops the `**/src/lib/queries.ts` max-lines exemption from
ui-dashboard/eslint.config.mjs — the barrel is ~6 lines and per-domain
files are well under the 1000-line hard cap. AGENTS.md §"File-size
budget" exemption note updated to match.

Existing module-wide snapshot + EXPECTED_EXPORT_NAMES test in
lib/__tests__/queries.test.ts pin all 7 query strings byte-for-byte
through the move and stay green without changes.
